# Files/swagger-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShortStay API",
    "version": "1.0.0",
    "description": "\n## ShortStay Backend API\n\nA full-featured short-term property rental platform API.\n\n### Authentication\nMost endpoints require a JWT Bearer token. Include it in the `Authorization` header:\n```\nAuthorization: Bearer <your_token>\n```\n\n### Roles\n- **guest** — Can search properties, make bookings, write reviews, raise complaints\n- **host** — Can list/manage properties, manage bookings, receive payouts\n- **admin** — Full system access: user management, property approval, reports\n- **accountant** — Manages payments, refunds, disputes, payouts\n- **verifier** — Conducts property inspections and submits reports\n\n### Google OAuth Flow\n1. Frontend redirects browser to `GET /api/auth/google`\n2. After Google login, backend redirects to `{FRONTEND_URL}/auth/callback?token=...&role=...&user_id=...`\n3. Frontend stores the token from the URL query parameter\n      ",
    "contact": {
      "name": "ShortStay Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token obtained from /api/auth/login"
      }
    },
...
  },
...
  "paths": {
...
    "/api/inspector/pending": {
      "get": {
        "summary": "Get all properties pending inspection (verifier only)",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of properties awaiting inspection"
          },
          "403": {
            "description": "Forbidden — verifier only"
          }
        }
      }
    },
    "/api/inspector/dashboard": {
      "get": {
        "summary": "Get inspector dashboard overview",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inspector dashboard stats (assigned, completed, pending)"
          },
          "403": {
            "description": "Forbidden — verifier only"
          }
        }
      }
    },
    "/api/inspector/history": {
      "get": {
        "summary": "Get inspection history for the logged-in inspector",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Past inspection records"
          },
          "403": {
            "description": "Forbidden — verifier only"
          }
        }
      }
    },
    "/api/inspector": {
      "get": {
        "summary": "Get all inspections assigned to the logged-in inspector",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of assigned inspections"
          },
          "403": {
            "description": "Forbidden — verifier only"
          }
        }
      }
    },
    "/api/inspector/submit": {
      "post": {
        "summary": "Submit an inspection report for a property",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "property_id",
                  "result",
                  "notes"
                ],
                "properties": {
                  "property_id": {
                    "type": "integer",
                    "example": 3
                  },
                  "result": {
                    "type": "string",
                    "enum": [
                      "passed",
                      "failed"
                    ],
                    "example": "passed"
                  },
                  "notes": {
                    "type": "string",
                    "example": "All safety standards met. Cleanliness is excellent."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inspection report submitted"
          },
          "403": {
            "description": "Forbidden — verifier only"
          }
        }
      }
    },
    "/api/inspector/badge/{property_id}": {
      "put": {
        "summary": "Approve a verified badge for a property",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Badge approved — property is_verified becomes true"
          },
          "403": {
            "description": "Forbidden — verifier only"
          }
        }
      }
    },
    "/api/inspector/{inspection_id}/images": {
      "post": {
        "summary": "Upload images for an inspection report (max 5)",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "inspection_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Inspection evidence images (max 5)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Images uploaded successfully"
          }
        }
      }
    },
    "/api/inspector/reports/inspections": {
      "get": {
        "summary": "Inspection report by date range (verifier only, ?format=csv supported)",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv"
              ]
            },
            "description": "Response format"
          }
        ],
        "responses": {
          "200": {
            "description": "Inspection report"
          }
        }
      }
    },
    "/api/inspector/reports/success-rate": {
      "get": {
        "summary": "Inspection pass/fail success rate report",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success rate statistics"
          }
        }
      }
    },
    "/api/inspector/reports/approved-vs-rejected": {
      "get": {
        "summary": "Approved vs rejected properties report",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Approved vs rejected comparison data"
          }
        }
      }
    },
    "/api/inspector/assign": {
      "post": {
        "summary": "Assign a verifier to a property (admin only)",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "property_id",
                  "inspector_id"
                ],
                "properties": {
                  "property_id": {
                    "type": "integer",
                    "example": 3
                  },
                  "inspector_id": {
                    "type": "integer",
                    "example": 7
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inspector assigned"
          },
          "403": {
            "description": "Forbidden — admins only"
          }
        }
      }
    },
    "/api/inspector/all": {
      "get": {
        "summary": "Get all inspections system-wide (admin only)",
        "tags": [
          "Inspector"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All inspection records"
          },
          "403": {
            "description": "Forbidden — admins only"
          }
        }
      }
    },
...
    "/api/dashboard/inspector": {
      "get": {
        "summary": "Get verifier dashboard (assigned inspections, completed count)",
        "tags": [
          "Dashboard"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inspector dashboard data"
          },
          "403": {
            "description": "Forbidden — verifier only"
          }
        }
      }
    },
...
  }
}